requirements.txt
flask>=2.1
flask-cors>=3.0
tensorflow>=2.12
opencv-python-headless>=4.8
numpy>=1.24
pillow>=10.0
scikit-learn>=1.2
pandas>=2.0
moviepy>=1.0
